+++ CHANGELOG.md
@@ -93,6 +93,10 @@
   ([\#5985](https://github.com/cometbft/cometbft/pull/5985))
 - `[types]` Update default max block bytes param to account for increased signature size of mldsa65.
   ([\#5987](https://github.com/cometbft/cometbft/pull/5987))
+- `[config]` Update the default max_tx_bytes to account for increased signature size of mlsdsa65.
+  ([\#5989](https://github.com/cometbft/cometbft/pull/5989))
+- `[crypto]` Add UnmarshalJSON to secp256k1eth key type.
+  ([\#5990](https://github.com/cometbft/cometbft/pull/5990))
 
 ### FEATURES